+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 1. Added composable to search weather
 2. Add user input to searh weather with latitude and longitude
+3. Add location entity for latitude and longitude